updated info for jobform and mutations
