+++ server/seeds/jobSeeds.json
@@ -2,6 +2,8 @@
   {
     "title": "Full Stack Developer",
     "description": "Junior position for JS",
+    "company": "Amazon",
+    "location": "Bend, Oregon",
     "link": "www.indeed.com",
     "skills": ["JavaScript", "Node.js"]
   },
